+++ src/features/advanced/index.ts
@@ -30,3 +30,4 @@ import './prod-hide-percent';
 import './prodq-hide-government-links';
 import './prodq-shorten-material-links';
 import './shpf-hide-sort-options';
+import './wf-workforce-filters/wf-workforce-filters';